Add glm-chaiscript
I found it super strange that nobody made bindings to glm publicly available, so I started working on these one. For now the principal functions for 3D vectors and quaternions are in, and the many other missing types are coming. I just started with what is the most useful to expose for my use-case (scripting for video game objects, with vector representing position, scale, euler angles and quaternion for absolute rotation/orientation)

As it is the goal of GLM anyway : function and type names mimic the ones of the GLSL language, the bindings are done without the "glm" namespace, so in chaiscript it will look like this : 

```chaiscript
var v = vec3(1.0, 2.0, 3.0);
```

Each defined type provide a "to_string" overload for chaiscript to use with "print"